Do not use cmake to set up test cases.
We used to use cmake to expand useful variables from the build system. For example:
@ILASM@, @TEST_CASE@ and so on.

Now we delegate that to the test case directive parser. We don't need to copy
the test case sources and result files in build folder anymore! We don't need to
track whether the source file has changed and call reconfigure. We don't need to 
implement very smart/difficult way to substitute the invocations for failed tests.

Simplify further the paths to the directories where the generated files will be placed.

PLEASE NOTE: Some of the test are still failing. This will be fixed in the next
few commits.