Implement delayed strong name signing.
We have a RSA signing of the assemblies. There is a public key which is uploaded
on the svn available to everyone, and before release they will be signed by the
private key which will be in some hidden place.

Note this required signing Cecil.Decompiler (one of our vendors) and its friends, too.