Remove the default constructors from the Shapes. Extend the arrow support.
Using default ctors of the shapes is evil because it forces the use of the 
Builder Pattern, which is *very* hard to debug. In the cases where it could be
avoided we should definitely do that.

Extend the variety of arrow heads and tails that we can draw. Documentation 
still to come.