Don't initialize the global properties. This overrides the settings if there was nested build system.
We reuse SolidOpt's build system in various of projects, in which SolidOpt is 
built as a vendor. Which means that when building the vendor it will initialize
those fields again and lose the information from the 'steering' build system.

This per-se is not a big deal for the vendor, however the 'real' project cannot
built proper solution, because of this missing/overriden information.