Fix the multiple regeneration of csprojs.
Usually we have a project which has several vendors. The vendors are configured and 
built by passing the control to the concrete vendor's build system. However the 
problem is that the concrete vendor's build system is not the 'steering' build 
system, and it doesn't know it.

We implement a lock, locking the generation (might be extended to lock more)
of cs projects by the 'steering' build system so that it will happen exactly one
time when all information is available.