List and refer csproj files within our automaticly generated csprojs.
So far we generated csproj files out of the cmake lists. However there are vendors
which come with csproj files. If that is the case we copy the csproj over in
our cmake build directory and set up some cmake variables in order to make the
vendor's csproj compatible with the current build.

For example if the build is by default for 2.0 our cmake will replace that with
the current cmake setup and so on.