Simplify the implementation of splitting and connecting blocks that contain exceptions.
Major issues was the semantics of finally and fault sections. In the end statically
we have to make them successors/predecessors because we don't know the concrete
control flow. (I.e in fault sections are different from finally: the code in 
the fault section is executed only when exception occurres)