[CMake] Move building nunit-console.exe to SolidOpt binary's folder.
nunit-console.exe needs dlls such as nunit-console-runner.dll, which is
is SolidOpt's build lib folder. If no MONO_PATH is specified the the
nunit-console couldn't be run and all tests fail. We add the lib folder
to the MONO_PATH environmental variable. The variable is used only
when the system runs the testsuite.